Refactor ChartJs into partial files
